LPD-91327 Move Tr*sted Hosts check from DCR service to DCR filter
Refactor only. The Tr*sted Hosts allowlist that LPD-XXXXX (commit
b9c12db68b) placed inside LiferayDynamicRegistrationService is moved
verbatim to DynamicRegistrationServiceContainerRequestFilter. No
behavior change, no new configuration, no test impact.

The original placement was a deliberate "consistency with the other
anonymous policies (scopes, redirect_uri patterns)" choice. Two
follow-up signals say it was the wrong cut:

1. The next gate to add to the anonymous DCR path is rate limiting,
   which conceptually belongs with the gates Liferay already runs in
   the filter ("can this caller reach the register endpoint at all?"
   - requireInitialAccessToken, soon rate limit) and not with the
   content policies the service runs ("given that this caller can
   reach the endpoint, what may they put in the registration body?"
   - allowed scopes, allowed redirect URI patterns). Tr*sted Hosts
   is also a gate by that criterion: it only looks at the client's
   IP, which is available before the JSON body is parsed.

2. Implementing rate limiting in the filter while leaving
   _validateAnonymousHosts in the service would require either (a)
   duplicating the _getClientHost helper in both places, or (b)
   extracting it to a shared util only to delete it the moment
   OAUTH2-33 lands and absorbs both helpers. Single source of truth
   wins; the helper lives where the gates that use it live.

The dividing line is now explicit: filter handles gates (decisions
based on the caller alone - IP, IAT presence, rate counters) before
the request body is parsed; the service handles policies (decisions
based on what the registration body asks for - scopes, redirect URIs)
after CXF has materialized ClientRegistration. The two anonymous
allowlists that remain in the service - anonymousAllowedScopes,
anonymousAllowedRedirectURIPatterns - are exactly the ones that
require the body.

Side effect that is intentional: the "Anonymous dynamic client
registration accepted" INFO log line now reports
_getClientHost(httpServletRequest) instead of
httpServletRequest.getRemoteAddr(). Behind a reverse proxy with
X-Forwarded-For, the log used to claim 127.0.0.1 while the policy
checked the real client IP; now both agree. Same logic, same
tr*st assumption that was implicit before, just consistent across
the two sites.

Verified post-deploy that the previous test matrix still holds:
- allowlist empty -> any IP passes;
- allowlist contains client's effective IP (as seen via XFF or TCP
  peer) -> 201;
- otherwise -> 403 with error_description naming the rejected IP.

Out of scope:
- The anonymousAllowedHosts configuration field itself (unchanged,
  same schema, same default of empty == no restriction).
- Translation files. The keys for the field were added by the
  Tr*sted Hosts commit and survive this refactor.
